Ejercicios 10/11/12

+++ src/Java/Ej3.java
@@ -4,12 +4,12 @@ public class Ej3 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Inserta un precio en la consola(Puede ser decimal)");
-        float num = sc.nextFloat();
-        final float iva = 21;
-        float valorAñadido = (num/100*iva);
-        float precioFinal = num+valorAñadido;
+        double num = sc.nextDouble();
+        final double iva = 21;
+        double valorAñadido = (num/100*iva);
+        double precioFinal = num+valorAñadido;
         System.out.println("El precio final será de : " + precioFinal);
-        //No entiendo porque cuando inserto un float salta error
+        //No entiendo porque cuando inserto un double salta error
     }
 
 }
